Split browser sign-in flow tests
