+++ Packages/Shared/CmuxAuthRuntime/Tests/CmuxAuthRuntimeTests/HostBrowserSignInFlowTests.swift
@@ -8,94 +8,13 @@ import Testing
 /// the sign-out-vs-callback race guards, deadlines, and attempt cancellation.
 @MainActor
 @Suite struct HostBrowserSignInFlowTests {
-    private struct Harness {
-        let flow: HostBrowserSignInFlow
-        let coordinator: AuthCoordinator
-        let client: FlowFakeAuthClient
-        let tokenStore: FlowInMemoryTokenStore
-        let factory: FakeBrowserAuthSessionFactory
-    }
-
-    private func makeHarness(
-        user: CMUXAuthUser? = nil,
-        browserAttemptTimeout: TimeInterval = 5 * 60,
-        slowSignInThreshold: TimeInterval = 30,
-        clock: (any Clock<Duration>)? = nil
-    ) -> Harness {
-        let store = FakeKeyValueStore()
-        // The fake client reads and clears the SAME token store the flow
-        // seeds, like production (StackAuthClient wraps the StackClientApp
-        // built over the store the callback seeds into). Split stores would
-        // hide races between the flow's seed handling and the coordinator's
-        // capture/clear sequence.
-        let tokenStore = FlowInMemoryTokenStore()
-        let client = FlowFakeAuthClient(user: user, store: tokenStore)
-        let coordinator = AuthCoordinator(
-            client: client,
-            sessionCache: CMUXAuthSessionCache(keyValueStore: store, key: "has_tokens"),
-            userCache: CMUXAuthIdentityStore(keyValueStore: store, key: "cached_user"),
-            teamSelection: CMUXAuthTeamSelectionStore(keyValueStore: store, key: "selected_team"),
-            anchor: FakeAnchor(),
-            config: .test,
-            launch: .plain()
-        )
-        let factory = FakeBrowserAuthSessionFactory()
-        let flow = HostBrowserSignInFlow(
-            coordinator: coordinator,
-            tokenStore: tokenStore,
-            sessionFactory: factory,
-            callbackRouter: AuthCallbackRouter(),
-            makeSignInURL: { URL(string: "https://example.test/handler/sign-in?cmux_auth_state=\($0)")! },
-            callbackScheme: { "cmux-dev" },
-            clock: clock ?? ContinuousClock(),
-            browserAttemptTimeout: browserAttemptTimeout,
-            slowSignInThreshold: slowSignInThreshold
-        )
-        return Harness(flow: flow, coordinator: coordinator, client: client, tokenStore: tokenStore, factory: factory)
-    }
-
-    private func callbackURL(state: String) -> URL {
-        URL(string: "cmux-dev://auth-callback?stack_refresh=refresh-1&stack_access=access-1&cmux_auth_state=\(state)")!
-    }
-
-    private func fallbackCallbackURL() -> URL {
-        URL(string: "cmux-dev://auth-callback?stack_refresh=refresh-1&stack_access=access-1")!
-    }
-
-    private func callbackState(_ session: FakeBrowserAuthSession) -> String {
-        URLComponents(url: session.signInURL, resolvingAgainstBaseURL: false)?
-            .queryItems?
-            .first(where: { $0.name == "cmux_auth_state" })?
-            .value ?? ""
-    }
-
-    private func waitForSession(_ factory: FakeBrowserAuthSessionFactory, count: Int = 1) async {
-        // The attempt task runs on the same main actor; yielding lets it reach
-        // the browser-session continuation deterministically.
-        while factory.sessions.count < count {
-            await Task.yield()
-        }
-    }
-
-    /// Yield to the main actor until `condition` holds. The work that flips the
-    /// condition (e.g. the attempt-timeout task resuming after the virtual clock
-    /// advances) runs on this same actor, so plain `Task.yield()` drains it with
-    /// no wall-clock dependence. Loops forever rather than failing on its own;
-    /// a stuck test surfaces as the suite-level timeout, matching the existing
-    /// `waitForSession` / `while harness.flow.isSigningIn` helpers in this file.
-    private func wait(until condition: @MainActor () -> Bool) async {
-        while !condition() {
-            await Task.yield()
-        }
-    }
-
     @Test func browserCallbackSignsInAndSeedsTokens() async throws {
         let user = CMUXAuthUser(id: "u1", primaryEmail: "a@b.com", displayName: "A")
-        let harness = makeHarness(user: user)
+        let harness = makeHostBrowserSignInFlowHarness(user: user)
 
         let attempt = Task { await harness.flow.signIn(timeout: 60) }
-        await waitForSession(harness.factory)
-        harness.factory.sessions[0].deliver(callbackURL(state: callbackState(harness.factory.sessions[0])))
+        await waitForHostBrowserSession(harness.factory)
+        harness.factory.sessions[0].deliver(hostBrowserCallbackURL(state: hostBrowserCallbackState(harness.factory.sessions[0])))
 
         #expect(await attempt.value)
         #expect(harness.coordinator.isAuthenticated)
@@ -105,63 +24,19 @@ import Testing
         #expect(harness.flow.isSigningIn == false)
     }
 
-    @Test func invalidCallbackPayloadIsRejected() async {
-        let harness = makeHarness(user: CMUXAuthUser(id: "u1", primaryEmail: nil, displayName: nil))
-
-        let attempt = Task { await harness.flow.signIn(timeout: 60) }
-        await waitForSession(harness.factory)
-        harness.factory.sessions[0].deliver(URL(string: "cmux-dev://auth-callback?other=1&cmux_auth_state=\(callbackState(harness.factory.sessions[0]))")!)
-
-        #expect(await attempt.value == false)
-        #expect(harness.coordinator.isAuthenticated == false)
-        #expect(await harness.tokenStore.getStoredRefreshToken() == nil)
-        #expect(harness.flow.lastFailure == .invalidCallback)
-    }
-
-    @Test func externalCallbackStateMismatchRecordsInvalidCallbackFailure() async {
-        let harness = makeHarness(user: CMUXAuthUser(id: "u1", primaryEmail: nil, displayName: nil))
-
-        let attempt = Task { await harness.flow.signIn(timeout: 60) }
-        await waitForSession(harness.factory)
-
-        let result = await harness.flow.handleCallbackURL(callbackURL(state: "other-state"))
-
-        #expect(result == false)
-        #expect(harness.flow.isSigningIn)
-        #expect(harness.coordinator.isAuthenticated == false)
-        #expect(harness.flow.lastFailure == .invalidCallback)
-
-        harness.factory.sessions[0].cancel()
-        #expect(await attempt.value == false)
-    }
-
-    @Test func callbackTokensThatDoNotValidateRecordUnauthorizedFailure() async {
-        let harness = makeHarness(user: nil)
-
-        let attempt = Task { await harness.flow.signIn(timeout: 60) }
-        await waitForSession(harness.factory)
-        harness.factory.sessions[0].deliver(callbackURL(state: callbackState(harness.factory.sessions[0])))
-
-        #expect(await attempt.value == false)
-        #expect(harness.coordinator.isAuthenticated == false)
-        #expect(await harness.tokenStore.getStoredRefreshToken() == "refresh-1")
-        #expect(await harness.tokenStore.getStoredAccessToken() == "access-1")
-        #expect(harness.flow.lastFailure == .unauthorized)
-    }
-
     @Test func nonAuthBrowserCompletionWaitsForExternalCallback() async {
         let user = CMUXAuthUser(id: "u1", primaryEmail: "a@b.com", displayName: "A")
-        let harness = makeHarness(user: user)
+        let harness = makeHostBrowserSignInFlowHarness(user: user)
 
         let attempt = Task { await harness.flow.signIn(timeout: 60) }
-        await waitForSession(harness.factory)
+        await waitForHostBrowserSession(harness.factory)
         harness.factory.sessions[0].deliver(URL(string: "https://example.test/handler/sign-in?after_auth_return_to=1")!)
 
         await Task.yield()
         #expect(harness.flow.isSigningIn)
         #expect(harness.coordinator.isAuthenticated == false)
 
-        let callbackResult = await harness.flow.handleCallbackURL(callbackURL(state: callbackState(harness.factory.sessions[0])))
+        let callbackResult = await harness.flow.handleCallbackURL(hostBrowserCallbackURL(state: hostBrowserCallbackState(harness.factory.sessions[0])))
 
         #expect(callbackResult)
         #expect(await attempt.value)
@@ -173,10 +48,10 @@ import Testing
     }
 
     @Test func cancelledPopupResolvesFalse() async {
-        let harness = makeHarness()
+        let harness = makeHostBrowserSignInFlowHarness()
 
         let attempt = Task { await harness.flow.signIn(timeout: 60) }
-        await waitForSession(harness.factory)
+        await waitForHostBrowserSession(harness.factory)
         harness.factory.sessions[0].cancel()
 
         #expect(await attempt.value == false)
@@ -184,10 +59,10 @@ import Testing
     }
 
     @Test func signOutCancelsActivePopup() async {
-        let harness = makeHarness()
+        let harness = makeHostBrowserSignInFlowHarness()
 
         harness.flow.beginSignIn()
-        await waitForSession(harness.factory)
+        await waitForHostBrowserSession(harness.factory)
         await harness.flow.signOut()
 
         #expect(harness.factory.sessions[0].cancelled)
@@ -196,12 +71,12 @@ import Testing
     }
 
     @Test func newAttemptCancelsPreviousPopup() async {
-        let harness = makeHarness()
+        let harness = makeHostBrowserSignInFlowHarness()
 
         harness.flow.beginSignIn()
-        await waitForSession(harness.factory)
+        await waitForHostBrowserSession(harness.factory)
         harness.flow.beginSignIn()
-        await waitForSession(harness.factory, count: 2)
+        await waitForHostBrowserSession(harness.factory, count: 2)
 
         #expect(harness.factory.sessions[0].cancelled)
         #expect(harness.factory.sessions[1].cancelled == false)
@@ -210,49 +85,26 @@ import Testing
 
     @Test func staleSessionCompletionCannotResumeNewAttempt() async {
         let user = CMUXAuthUser(id: "u1", primaryEmail: "a@b.com", displayName: "A")
-        let harness = makeHarness(user: user)
+        let harness = makeHostBrowserSignInFlowHarness(user: user)
 
         harness.flow.beginSignIn()
-        await waitForSession(harness.factory)
+        await waitForHostBrowserSession(harness.factory)
         let staleSession = harness.factory.sessions[0]
         staleSession.deliverCancelCompletion = false
 
         let secondAttempt = Task { await harness.flow.signIn(timeout: 60) }
-        await waitForSession(harness.factory, count: 2)
-        staleSession.deliver(callbackURL(state: callbackState(staleSession)))
+        await waitForHostBrowserSession(harness.factory, count: 2)
+        staleSession.deliver(hostBrowserCallbackURL(state: hostBrowserCallbackState(staleSession)))
 
         await Task.yield()
         #expect(harness.flow.isSigningIn)
         #expect(harness.coordinator.isAuthenticated == false)
 
-        harness.factory.sessions[1].deliver(callbackURL(state: callbackState(harness.factory.sessions[1])))
+        harness.factory.sessions[1].deliver(hostBrowserCallbackURL(state: hostBrowserCallbackState(harness.factory.sessions[1])))
 
         #expect(await secondAttempt.value)
         #expect(harness.coordinator.isAuthenticated)
         #expect(harness.coordinator.currentUser == user)
-    }
-
-    @Test func abandonedBrowserAttemptTimesOut() async throws {
-        // Drive the abandoned-attempt timeout off a virtual clock so the result
-        // does not depend on a real-timer task being scheduled within a fixed
-        // wall-clock window. beginSignIn parks two sleepers on this clock: the
-        // attempt timeout (1s here) and the slow-sign-in hint (30s default).
-        let clock = ManualTestClock()
-        let harness = makeHarness(browserAttemptTimeout: 1, clock: clock)
-
-        harness.flow.beginSignIn()
-        await waitForSession(harness.factory)
-        await clock.waitUntilSleepers(count: 2)
-
-        // Advance past the attempt timeout but well under the 30s slow-hint
-        // threshold, so only the abandoned-attempt timeout fires.
-        clock.advance(by: .seconds(1))
-
-        await wait { harness.flow.isSigningIn == false }
-        #expect(harness.factory.sessions[0].cancelled)
-        #expect(harness.flow.isSigningIn == false)
-        #expect(harness.coordinator.isAuthenticated == false)
-        #expect(harness.flow.lastFailure == .timedOut)
     }
 
     @Test func slowSignInSurfacesBrowserFallback() async throws {
@@ -262,11 +114,11 @@ import Testing
         // staring at a dead window. Past the slow threshold the flow must flip
         // `signInIsSlow` so the account UI can offer the "open in your default
         // browser" fallback instead of an indefinite spinner.
-        let harness = makeHarness(slowSignInThreshold: 0.05)
+        let harness = makeHostBrowserSignInFlowHarness(slowSignInThreshold: 0.05)
         #expect(harness.flow.signInIsSlow == false)
 
         harness.flow.beginSignIn()
-        await waitForSession(harness.factory)
+        await waitForHostBrowserSession(harness.factory)
 
         var becameSlow = false
         for _ in 0..<200 {
@@ -290,11 +142,11 @@ import Testing
     }
 
     @Test func activeAttemptSignInURLCarriesActiveAttemptState() async {
-        let harness = makeHarness()
+        let harness = makeHostBrowserSignInFlowHarness()
         #expect(harness.flow.activeAttemptSignInURL == nil)
 
         harness.flow.beginSignIn()
-        await waitForSession(harness.factory)
+        await waitForHostBrowserSession(harness.factory)
 
         let fallbackURL = harness.flow.activeAttemptSignInURL
         #expect(fallbackURL != nil)
@@ -307,15 +159,15 @@ import Testing
                 .first(where: { $0.name == "cmux_auth_state" })?
                 .value
         }
-        #expect(fallbackState == callbackState(harness.factory.sessions[0]))
+        #expect(fallbackState == hostBrowserCallbackState(harness.factory.sessions[0]))
     }
 
     @Test func issuedFallbackCallbackSurvivesPopupCancellation() async throws {
         let user = CMUXAuthUser(id: "u1", primaryEmail: "a@b.com", displayName: "A")
-        let harness = makeHarness(user: user)
+        let harness = makeHostBrowserSignInFlowHarness(user: user)
 
         harness.flow.beginSignIn()
-        await waitForSession(harness.factory)
+        await waitForHostBrowserSession(harness.factory)
         let fallbackURL = try #require(harness.flow.activeAttemptSignInURL)
         let fallbackState = try #require(URLComponents(url: fallbackURL, resolvingAgainstBaseURL: false)?
             .queryItems?
@@ -327,7 +179,7 @@ import Testing
             await Task.yield()
         }
 
-        let callbackResult = await harness.flow.handleCallbackURL(callbackURL(state: fallbackState))
+        let callbackResult = await harness.flow.handleCallbackURL(hostBrowserCallbackURL(state: fallbackState))
 
         #expect(callbackResult)
         #expect(harness.coordinator.isAuthenticated)
@@ -338,10 +190,10 @@ import Testing
 
     @Test func issuedFallbackCallbackAfterSignOutIsRejected() async throws {
         let user = CMUXAuthUser(id: "u1", primaryEmail: "a@b.com", displayName: "A")
-        let harness = makeHarness(user: user)
+        let harness = makeHostBrowserSignInFlowHarness(user: user)
 
         harness.flow.beginSignIn()
-        await waitForSession(harness.factory)
+        await waitForHostBrowserSession(harness.factory)
         let fallbackURL = try #require(harness.flow.activeAttemptSignInURL)
         let fallbackState = try #require(URLComponents(url: fallbackURL, resolvingAgainstBaseURL: false)?
             .queryItems?
@@ -354,7 +206,7 @@ import Testing
         }
         await harness.flow.signOut()
 
-        let callbackResult = await harness.flow.handleCallbackURL(callbackURL(state: fallbackState))
+        let callbackResult = await harness.flow.handleCallbackURL(hostBrowserCallbackURL(state: fallbackState))
 
         #expect(callbackResult == false)
         #expect(harness.coordinator.isAuthenticated == false)
@@ -364,12 +216,12 @@ import Testing
 
     @Test func signOutDuringCallbackValidationWins() async {
         let user = CMUXAuthUser(id: "u1", primaryEmail: "a@b.com", displayName: "A")
-        let harness = makeHarness(user: user)
+        let harness = makeHostBrowserSignInFlowHarness(user: user)
         await harness.client.closeUserGate()
 
         let attempt = Task { await harness.flow.signIn(timeout: 60) }
-        await waitForSession(harness.factory)
-        harness.factory.sessions[0].deliver(callbackURL(state: callbackState(harness.factory.sessions[0])))
+        await waitForHostBrowserSession(harness.factory)
+        harness.factory.sessions[0].deliver(hostBrowserCallbackURL(state: hostBrowserCallbackState(harness.factory.sessions[0])))
         // Wait until the completion path is blocked inside the user fetch.
         while await harness.client.pendingUserRequests == 0 {
             await Task.yield()
@@ -394,12 +246,12 @@ import Testing
         // (the callback path already cancelled the timeout before parking).
         let clock = ManualTestClock()
         let user = CMUXAuthUser(id: "u1", primaryEmail: "a@b.com", displayName: "A")
-        let harness = makeHarness(user: user, browserAttemptTimeout: 1, clock: clock)
+        let harness = makeHostBrowserSignInFlowHarness(user: user, browserAttemptTimeout: 1, clock: clock)
         await harness.client.closeUserGate()
 
         let attempt = Task { await harness.flow.signIn(timeout: 60) }
-        await waitForSession(harness.factory)
-        harness.factory.sessions[0].deliver(callbackURL(state: callbackState(harness.factory.sessions[0])))
+        await waitForHostBrowserSession(harness.factory)
+        harness.factory.sessions[0].deliver(hostBrowserCallbackURL(state: hostBrowserCallbackState(harness.factory.sessions[0])))
         while await harness.client.pendingUserRequests == 0 {
             await Task.yield()
         }
@@ -419,7 +271,7 @@ import Testing
 
     @Test func deadlineResolvesFalseWhilePopupStaysUp() async {
         let user = CMUXAuthUser(id: "u1", primaryEmail: "a@b.com", displayName: "A")
-        let harness = makeHarness(user: user)
+        let harness = makeHostBrowserSignInFlowHarness(user: user)
 
         let result = await harness.flow.signIn(timeout: 0.05)
         #expect(result == false)
@@ -427,7 +279,7 @@ import Testing
         #expect(harness.factory.sessions[0].cancelled == false)
 
         // The user can still finish in the popup after the caller's deadline.
-        harness.factory.sessions[0].deliver(callbackURL(state: callbackState(harness.factory.sessions[0])))
+        harness.factory.sessions[0].deliver(hostBrowserCallbackURL(state: hostBrowserCallbackState(harness.factory.sessions[0])))
         while harness.coordinator.isAuthenticated == false {
             await Task.yield()
         }
@@ -436,11 +288,11 @@ import Testing
 
     @Test func lateCallbackAfterSignOutIsRejected() async {
         let user = CMUXAuthUser(id: "u1", primaryEmail: "a@b.com", displayName: "A")
-        let harness = makeHarness(user: user)
+        let harness = makeHostBrowserSignInFlowHarness(user: user)
 
         harness.flow.beginSignIn()
-        await waitForSession(harness.factory)
-        let staleCallback = callbackURL(state: callbackState(harness.factory.sessions[0]))
+        await waitForHostBrowserSession(harness.factory)
+        let staleCallback = hostBrowserCallbackURL(state: hostBrowserCallbackState(harness.factory.sessions[0]))
         await harness.flow.signOut()
 
         let result = await harness.flow.handleCallbackURL(staleCallback)
@@ -453,11 +305,11 @@ import Testing
 
     @Test func mismatchedCallbackStateIsRejected() async {
         let user = CMUXAuthUser(id: "u1", primaryEmail: "a@b.com", displayName: "A")
-        let harness = makeHarness(user: user)
+        let harness = makeHostBrowserSignInFlowHarness(user: user)
 
         let attempt = Task { await harness.flow.signIn(timeout: 60) }
-        await waitForSession(harness.factory)
-        harness.factory.sessions[0].deliver(callbackURL(state: "stale-state"))
+        await waitForHostBrowserSession(harness.factory)
+        harness.factory.sessions[0].deliver(hostBrowserCallbackURL(state: "stale-state"))
 
         #expect(await attempt.value == false)
         #expect(harness.coordinator.isAuthenticated == false)
@@ -466,18 +318,18 @@ import Testing
 
     @Test func staleExternalCallbackDoesNotCancelActiveAttempt() async {
         let user = CMUXAuthUser(id: "u1", primaryEmail: "a@b.com", displayName: "A")
-        let harness = makeHarness(user: user)
+        let harness = makeHostBrowserSignInFlowHarness(user: user)
 
         let attempt = Task { await harness.flow.signIn(timeout: 60) }
-        await waitForSession(harness.factory)
+        await waitForHostBrowserSession(harness.factory)
 
-        let staleResult = await harness.flow.handleCallbackURL(callbackURL(state: "stale-state"))
+        let staleResult = await harness.flow.handleCallbackURL(hostBrowserCallbackURL(state: "stale-state"))
 
         #expect(staleResult == false)
         #expect(harness.flow.isSigningIn)
         #expect(harness.coordinator.isAuthenticated == false)
 
-        harness.factory.sessions[0].deliver(callbackURL(state: callbackState(harness.factory.sessions[0])))
+        harness.factory.sessions[0].deliver(hostBrowserCallbackURL(state: hostBrowserCallbackState(harness.factory.sessions[0])))
 
         #expect(await attempt.value)
         #expect(harness.coordinator.isAuthenticated)
@@ -487,9 +339,9 @@ import Testing
 
     @Test func fallbackExternalCallbackWithoutActiveAttemptSignsIn() async {
         let user = CMUXAuthUser(id: "u1", primaryEmail: "a@b.com", displayName: "A")
-        let harness = makeHarness(user: user)
+        let harness = makeHostBrowserSignInFlowHarness(user: user)
 
-        let result = await harness.flow.handleCallbackURL(fallbackCallbackURL())
+        let result = await harness.flow.handleCallbackURL(hostBrowserFallbackCallbackURL())
 
         #expect(result)
         #expect(harness.coordinator.isAuthenticated)
@@ -500,9 +352,9 @@ import Testing
 
     @Test func statefulExternalCallbackWithoutActiveAttemptIsRejected() async {
         let user = CMUXAuthUser(id: "u1", primaryEmail: "a@b.com", displayName: "A")
-        let harness = makeHarness(user: user)
+        let harness = makeHostBrowserSignInFlowHarness(user: user)
 
-        let result = await harness.flow.handleCallbackURL(callbackURL(state: "stale-state"))
+        let result = await harness.flow.handleCallbackURL(hostBrowserCallbackURL(state: "stale-state"))
 
         #expect(result == false)
         #expect(harness.coordinator.isAuthenticated == false)
@@ -520,12 +372,12 @@ import Testing
         // device is online. The coordinator owns the local clear AFTER the
         // capture; the flow must not clear the shared store underneath it.
         let user = CMUXAuthUser(id: "u1", primaryEmail: "a@b.com", displayName: "A")
-        let harness = makeHarness(user: user)
+        let harness = makeHostBrowserSignInFlowHarness(user: user)
         await harness.client.closeUserGate()
 
         let attempt = Task { await harness.flow.signIn(timeout: 60) }
-        await waitForSession(harness.factory)
-        harness.factory.sessions[0].deliver(callbackURL(state: callbackState(harness.factory.sessions[0])))
+        await waitForHostBrowserSession(harness.factory)
+        harness.factory.sessions[0].deliver(hostBrowserCallbackURL(state: hostBrowserCallbackState(harness.factory.sessions[0])))
         while await harness.client.pendingUserRequests == 0 {
             await Task.yield()
         }
